More cleanup to add flutter UI to sub projects

+++ sources/sub-projects/ayfstarter/post_install/install_flutter_app.sh
@@ -28,16 +28,22 @@ echo "Running zip release with path=$PATH"
 echo "flutter build $2"
 flutter build $2
 
-echo cp $1/.$3.env build/$sysfolder/runner/Release/data/flutter_assets
-cp $1/.$3.env build/$sysfolder/runner/Release/data/flutter_assets
+if [[ "$2" == "web" ]]
+then
+	# What to do here??
+	echo "Compiling for target web"
+else
+	echo cp $1/.$3.env build/$sysfolder/runner/Release/data/flutter_assets
+	cp $1/.$3.env build/$sysfolder/runner/Release/data/flutter_assets
 
-if [ -d "$4/$3" ]; then
-	echo "rm -rf $4/$3"
-	rm -rf $4/$3
+	if [ -d "$4/$3" ]; then
+		echo "rm -rf $4/$3"
+		rm -rf $4/$3
+	fi
+
+	echo "mv build/$sysfolder/runner/Release $4/$3"
+	mv build/$sysfolder/runner/Release $4/$3
 fi
-
-echo "mv build/$sysfolder/runner/Release $4/$3"
-mv build/$sysfolder/runner/Release $4/$3
 
 echo "Flutter app preparation completed!"
